refactor : FrameTest를 NormalFrameTest와 FinalFrameTest로 분리
